// lib/generated/prisma-node/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  DEPARTMENT_ADMIN: 'DEPARTMENT_ADMIN',
  LECTURER: 'LECTURER',
  STUDENT: 'STUDENT',
  OFFICE_STAFF: 'OFFICE_STAFF'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Permission = {
  VIEW: 'VIEW',
  DOWNLOAD: 'DOWNLOAD',
  EDIT: 'EDIT'
} as const

export type Permission = (typeof Permission)[keyof typeof Permission]


export const ShareTargetType = {
  USER: 'USER',
  OFFICE: 'OFFICE',
  GROUP: 'GROUP',
  DEPARTMENT: 'DEPARTMENT'
} as const

export type ShareTargetType = (typeof ShareTargetType)[keyof typeof ShareTargetType]


export const DocumentType = {
  NOTE: 'NOTE',
  SHARED_FILE: 'SHARED_FILE'
} as const

export type DocumentType = (typeof DocumentType)[keyof typeof DocumentType]


export const StaffTokenCategory = {
  LECTURER: 'LECTURER',
  HOD_DEAN: 'HOD_DEAN',
  HIGHER_AUTHORITY: 'HIGHER_AUTHORITY'
} as const

export type StaffTokenCategory = (typeof StaffTokenCategory)[keyof typeof StaffTokenCategory]


export const SupervisionStatus = {
  ACTIVE: 'ACTIVE',
  COMPLETED: 'COMPLETED',
  ON_HOLD: 'ON_HOLD'
} as const

export type SupervisionStatus = (typeof SupervisionStatus)[keyof typeof SupervisionStatus]


export const MentorStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type MentorStatus = (typeof MentorStatus)[keyof typeof MentorStatus]


export const LostFoundType = {
  MISSING: 'MISSING',
  FOUND: 'FOUND'
} as const

export type LostFoundType = (typeof LostFoundType)[keyof typeof LostFoundType]